api 주소 변경 및 칵테일 리뷰 수정 삭제

+++ apis/review/reviewResolver.py
@@ -14,14 +14,14 @@ class Create_review(Resource):
         # try:
         # logined_user = User.query.filter_by(email=session['login']).first()
         # user_id = logined_user.id
-        # if not session:
-        #     abort(500, "로그인 해주세요")
-        # else:
-        user_id = request.json['user_id']
-        liquor_id = request.json['liquor_id']
-        rating = request.json['rating']
-        content = request.json['content']
-        return reviewService.create_review(user_id,liquor_id,rating,content)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            user_id = request.json['user_id']
+            liquor_id = request.json['liquor_id']
+            rating = request.json['rating']
+            content = request.json['content']
+            return reviewService.create_review(user_id,liquor_id,rating,content)
         # except:
         #     abort(500, "리뷰 등록 실패.")
 
@@ -32,21 +32,21 @@ class UpdateReview(Resource):
     @Review.response(500, 'Failed to revise the review')
     def post(self):
         '''술 리뷰 수정'''
-        # if not session:
-        #     abort(500, "로그인 해주세요")
-        # else:
-        user_id = request.json['user_id']
-        liquor_id = request.json['liquor_id']
-        rating = request.json['rating']
-        content = request.json['content']
-        return reviewService.update_review(user_id,liquor_id,rating,content)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            user_id = request.json['user_id']
+            liquor_id = request.json['liquor_id']
+            rating = request.json['rating']
+            content = request.json['content']
+            return reviewService.update_review(user_id,liquor_id,rating,content)
         # except:
         #     abort(500, "로그인 해주세요.")
 
-@Review.route("/delete/user_id=<int:user_id>/liouor_id=<int:liquor_id>")
+@Review.route("/delete/<int:user_id>/liquor/<int:liquor_id>")
 class DeleteReview(Resource):
     def delete(self, user_id, liquor_id):
-    #     if not session:
-    #         abort(500, "로그인 해주세요")
-    #     else:
-        return reviewService.delete_review( user_id, liquor_id)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            return reviewService.delete_review( user_id, liquor_id)
